move hover and focus to actout
