+++ src/themes/greyVest.js
@@ -2,7 +2,8 @@ import React from 'react'
 import _ from 'lodash/fp'
 import * as F from 'futil-js'
 import { observer } from 'mobx-react'
-import { partial, withStateLens, hover } from '../utils/mobx-react-utils'
+import { hover } from '../utils/actout'
+import { partial, withStateLens } from '../utils/mobx-react-utils'
 import { Flex, TextHighlight, FilteredPicker, ModalFilterAdder } from '../'
 import ExampleTypeConstructor from '../exampleTypes/'
